Split text rendering part of pal into its own class, QgsTextRenderer
