+++ python/core/qgspallabeling.sip
@@ -77,7 +77,7 @@
 %End
 };
 
-class QgsPalLayerSettings
+class QgsPalLayerSettings : QgsTextRendererSettings
 {
 %TypeHeaderCode
 #include <qgspallabeling.h>
@@ -134,53 +134,6 @@ class QgsPalLayerSettings
       SymbolLeftRight, // place direction symbols on left/right of label
       SymbolAbove, // place direction symbols on above label
       SymbolBelow // place direction symbols on below label
-    };
-
-    enum MultiLineAlign
-    {
-      MultiLeft,
-      MultiCenter,
-      MultiRight
-    };
-
-    enum ShapeType
-    {
-      ShapeRectangle,
-      ShapeSquare,
-      ShapeEllipse,
-      ShapeCircle,
-      ShapeSVG
-    };
-
-    enum SizeType
-    {
-      SizeBuffer,
-      SizeFixed,
-      SizePercent
-    };
-
-    enum RotationType
-    {
-      RotationSync,
-      RotationOffset,
-      RotationFixed
-    };
-
-    /** Units used for option sizes, before being converted to rendered sizes */
-    enum SizeUnit
-    {
-      Points,
-      MM,
-      MapUnits,
-      Percent
-    };
-
-    enum ShadowType
-    {
-      ShadowLowest,
-      ShadowText,
-      ShadowBuffer,
-      ShadowShape
     };
 
     // update mDataDefinedNames QMap in constructor when adding/deleting enum value
@@ -304,20 +257,7 @@ class QgsPalLayerSettings
       */
     QgsExpression* getLabelExpression();
 
-    QFont textFont;
-    QString textNamedStyle;
-    bool fontSizeInMapUnits; //true if font size is in map units (otherwise in points)
-    QgsMapUnitScale fontSizeMapUnitScale; // scale range for map units for font size
-    QColor textColor;
-    int textTransp;
-    QPainter::CompositionMode blendMode;
-    QColor previewBkgrdColor;
-
     //-- text formatting
-
-    QString wrapChar;
-    double multilineHeight; //0.0 to 10.0, leading between lines as multiplyer of line height
-    MultiLineAlign multilineAlign; // horizontal alignment of multi-line labels
 
     // Adds '<' or '>', or user-defined symbol to the label string pointing to the
     // direction of the line / polygon ring
@@ -331,62 +271,6 @@ class QgsPalLayerSettings
     bool formatNumbers;
     int decimals;
     bool plusSign;
-
-    //-- text buffer
-
-    bool bufferDraw;
-    double bufferSize; // buffer size
-    bool bufferSizeInMapUnits; //true if buffer is in map units (otherwise in mm)
-    QgsMapUnitScale bufferSizeMapUnitScale; // scale range for map units for buffer size
-    QColor bufferColor;
-    bool bufferNoFill; //set interior of buffer to 100% transparent
-    int bufferTransp;
-    Qt::PenJoinStyle bufferJoinStyle;
-    QPainter::CompositionMode bufferBlendMode;
-
-    //-- shape background
-
-    bool shapeDraw;
-    ShapeType shapeType;
-    QString shapeSVGFile;
-    SizeType shapeSizeType;
-    QPointF shapeSize;
-    SizeUnit shapeSizeUnits;
-    QgsMapUnitScale shapeSizeMapUnitScale;
-    RotationType shapeRotationType;
-    double shapeRotation;
-    QPointF shapeOffset;
-    SizeUnit shapeOffsetUnits;
-    QgsMapUnitScale shapeOffsetMapUnitScale;
-    QPointF shapeRadii;
-    SizeUnit shapeRadiiUnits;
-    QgsMapUnitScale shapeRadiiMapUnitScale;
-    int shapeTransparency;
-    QPainter::CompositionMode shapeBlendMode;
-    QColor shapeFillColor;
-    QColor shapeBorderColor;
-    double shapeBorderWidth;
-    SizeUnit shapeBorderWidthUnits;
-    QgsMapUnitScale shapeBorderWidthMapUnitScale;
-    Qt::PenJoinStyle shapeJoinStyle;
-
-    //-- drop shadow
-
-    bool shadowDraw;
-    ShadowType shadowUnder;
-    int shadowOffsetAngle;
-    double shadowOffsetDist;
-    SizeUnit shadowOffsetUnits;
-    QgsMapUnitScale shadowOffsetMapUnitScale;
-    bool shadowOffsetGlobal;
-    double shadowRadius;
-    SizeUnit shadowRadiusUnits;
-    QgsMapUnitScale shadowRadiusMapUnitScale;
-    bool shadowRadiusAlphaOnly;
-    int shadowTransparency;
-    int shadowScale;
-    QColor shadowColor;
-    QPainter::CompositionMode shadowBlendMode;
 
     //-- placement
 
@@ -440,10 +324,6 @@ class QgsPalLayerSettings
 
     double minFeatureSize; // minimum feature size to be labelled (in mm)
     bool obstacle; // whether features for layer are obstacles to labels of other layers
-
-    //-- scale factors
-    double vectorScaleFactor; //scale factor painter units->pixels
-    double rasterCompressFactor; //pixel resolution scale factor
 
     // called from register feature hook
     void calculateLabelSize( const QFontMetricsF* fm, QString text, double& labelX, double& labelY, QgsFeature* f = 0 );
@@ -513,17 +393,6 @@ class QgsPalLayerSettings
      */
     int sizeToPixel( double size, const QgsRenderContext& c , SizeUnit unit, bool rasterfactor = false, const QgsMapUnitScale& mapUnitScale = QgsMapUnitScale() ) const;
 
-    /** Calculates size (considering output size should be in pixel or map units, scale factors and optionally oversampling)
-     * @param size size to convert
-     * @param c rendercontext
-     * @param unit SizeUnit enum value of size
-     * @param rasterfactor whether to consider oversampling
-     * @param mapUnitScale a mapUnitScale clamper
-     * @return size that will render, as double
-     * @note added in 1.9, as a better precision replacement for sizeToPixel
-     */
-    double scaleToPixelContext( double size, const QgsRenderContext& c, SizeUnit unit, bool rasterfactor = false, const QgsMapUnitScale& mapUnitScale = QgsMapUnitScale() ) const;
-
     /** Map of data defined enum to names and old-style indecies
      * The QPair contains a new string for layer property key, and a reference to old-style numeric key (< QGIS 2.0)
      * @note not available in python bindings; added in 1.9
@@ -543,62 +412,6 @@ class QgsLabelCandidate
     QRectF rect;
     double cost;
 };
-
-/** \ingroup core
-  * Maintains current state of more grainular and temporal values when creating/painting
-  * component parts of an individual label (e.g. buffer, background, shadow, etc.).
-  */
-class QgsLabelComponent
-{
-%TypeHeaderCode
-#include <qgspallabeling.h>
-%End
-
-  public:
-    QgsLabelComponent();
-
-    // methods
-
-    const QString& text();
-    void setText( const QString& text );
-
-    const QgsPoint& origin();
-    void setOrigin( QgsPoint point );
-
-    bool useOrigin() const;
-    void setUseOrigin( bool use );
-
-    double rotation() const;
-    void setRotation( double rotation );
-
-    double rotationOffset() const;
-    void setRotationOffset( double rotation );
-
-    bool useRotation() const;
-    void setUseRotation( bool use );
-
-    const QgsPoint& center();
-    void setCenter( QgsPoint point );
-
-    bool useCenter() const;
-    void setUseCenter( bool use );
-
-    const QgsPoint& size();
-    void setSize( QgsPoint point );
-
-    const QgsPoint& offset();
-    void setOffset( QgsPoint point );
-
-    const QPicture* picture();
-    void setPicture( QPicture* picture );
-
-    double pictureBuffer() const;
-    void setPictureBuffer( double buffer );
-
-    double dpiRatio() const;
-    void setDpiRatio( double ratio );
-};
-
 
 /**
  * Class that stores computed placement from labeling engine.
@@ -631,14 +444,6 @@ class QgsPalLabeling : QgsLabelingEngineInterface
 %End
 
   public:
-    enum DrawLabelType
-    {
-      LabelText,
-      LabelBuffer,
-      LabelShape,
-      LabelSVG,
-      LabelShadow
-    };
 
     QgsPalLabeling();
     ~QgsPalLabeling();
