+++ src/core/dxf/qgsdxfpallabeling.h
@@ -19,6 +19,7 @@
 #define QGSDXFPALLABELING_H
 
 #include "qgspallabeling.h"
+#include "qgstextrenderer.h"
 #include "qgsmaprenderer.h"
 #include "qgsrendercontext.h"
 
@@ -31,7 +32,7 @@ class CORE_EXPORT QgsDxfPalLabeling: public QgsPalLabeling
     ~QgsDxfPalLabeling();
 
     QgsRenderContext& renderContext() { return mRenderContext; }
-    void drawLabel( pal::LabelPosition* label, QgsRenderContext& context, QgsPalLayerSettings& tmpLyr, DrawLabelType drawType, double dpiRatio = 1.0 );
+    void drawLabel( pal::LabelPosition* label, QgsRenderContext& context, QgsPalLayerSettings& tmpLyr, QgsTextRenderer::TextComponentPart drawType, double dpiRatio = 1.0 );
 
   private:
     QgsDxfExport* mDxfExport;
