+++ src/core/qgspallabeling.cpp
@@ -58,21 +58,6 @@
 #include "qgssymbollayerv2utils.h"
 #include <QMessageBox>
 
-
-Q_GUI_EXPORT extern int qt_defaultDpiX();
-Q_GUI_EXPORT extern int qt_defaultDpiY();
-
-static void _fixQPictureDPI( QPainter* p )
-{
-  // QPicture makes an assumption that we drawing to it with system DPI.
-  // Then when being drawn, it scales the painter. The following call
-  // negates the effect. There is no way of setting QPicture's DPI.
-  // See QTBUG-20361
-  p->scale(( double )qt_defaultDpiX() / p->device()->logicalDpiX(),
-           ( double )qt_defaultDpiY() / p->device()->logicalDpiY() );
-}
-
-
 using namespace pal;
 
 #if 0
@@ -224,10 +209,10 @@ class QgsPalGeometry : public PalGeometry
 // -------------
 
 QgsPalLayerSettings::QgsPalLayerSettings()
-    : palLayer( NULL )
+    : QgsTextRendererSettings()
+    , palLayer( NULL )
     , mCurFeat( 0 )
     , mCurFields( 0 )
-    , ct( NULL )
     , extentGeom( NULL )
     , mFeaturesToLabel( 0 )
     , mFeatsSendingToPal( 0 )
@@ -236,22 +221,7 @@ QgsPalLayerSettings::QgsPalLayerSettings()
 {
   enabled = false;
 
-  // text style
-  textFont = QApplication::font();
-  textNamedStyle = QString( "" );
-  fontSizeInMapUnits = false;
-  textColor = Qt::black;
-  textTransp = 0;
-  blendMode = QPainter::CompositionMode_SourceOver;
-  previewBkgrdColor = Qt::white;
-  // font processing info
-  mTextFontFound = true;
-  mTextFontFamily = QApplication::font().family();
-
   // text formatting
-  wrapChar = "";
-  multilineHeight = 1.0;
-  multilineAlign = MultiLeft;
   addDirectionSymbol = false;
   leftDirectionSymbol = QString( "<" );
   rightDirectionSymbol = QString( ">" );
@@ -260,52 +230,6 @@ QgsPalLayerSettings::QgsPalLayerSettings()
   formatNumbers = false;
   decimals = 3;
   plusSign = false;
-
-  // text buffer
-  bufferDraw = false;
-  bufferSize = 1.0;
-  bufferSizeInMapUnits = false;
-  bufferColor = Qt::white;
-  bufferTransp = 0;
-  bufferNoFill = false;
-  bufferJoinStyle = Qt::BevelJoin;
-  bufferBlendMode = QPainter::CompositionMode_SourceOver;
-
-  // shape background
-  shapeDraw = false;
-  shapeType = ShapeRectangle;
-  shapeSVGFile = QString();
-  shapeSizeType = SizeBuffer;
-  shapeSize = QPointF( 0.0, 0.0 );
-  shapeSizeUnits = MM;
-  shapeRotationType = RotationSync;
-  shapeRotation = 0.0;
-  shapeOffset = QPointF( 0.0, 0.0 );
-  shapeOffsetUnits = MM;
-  shapeRadii = QPointF( 0.0, 0.0 );
-  shapeRadiiUnits = MM;
-  shapeFillColor = Qt::white;
-  shapeBorderColor = Qt::darkGray;
-  shapeBorderWidth = 0.0;
-  shapeBorderWidthUnits = MM;
-  shapeJoinStyle = Qt::BevelJoin;
-  shapeTransparency = 0;
-  shapeBlendMode = QPainter::CompositionMode_SourceOver;
-
-  // drop shadow
-  shadowDraw = false;
-  shadowUnder = ShadowLowest;
-  shadowOffsetAngle = 135;
-  shadowOffsetDist = 1.0;
-  shadowOffsetUnits = MM;
-  shadowOffsetGlobal = true;
-  shadowRadius = 1.5;
-  shadowRadiusUnits = MM;
-  shadowRadiusAlphaOnly = false;
-  shadowTransparency = 30;
-  shadowScale = 100;
-  shadowColor = Qt::black;
-  shadowBlendMode = QPainter::CompositionMode_Multiply;
 
   // placement
   placement = AroundPoint;
@@ -342,10 +266,6 @@ QgsPalLayerSettings::QgsPalLayerSettings()
   limitNumLabels = false;
   maxNumLabels = 2000;
   obstacle = true;
-
-  // scale factors
-  vectorScaleFactor = 1.0;
-  rasterCompressFactor = 1.0;
 
   // data defined string and old-style index values
   // NOTE: in QPair use -1 for second value (other values are for old-style layer properties migration)
@@ -452,11 +372,10 @@ QgsPalLayerSettings::QgsPalLayerSettings()
   mDataDefinedNames.insert( Show, QPair<QString, int>( "Show", 15 ) );
   mDataDefinedNames.insert( AlwaysShow, QPair<QString, int>( "AlwaysShow", 20 ) );
 
-  // temp stuff for when drawing label components (don't copy)
-  showingShadowRects = false;
 }
 
 QgsPalLayerSettings::QgsPalLayerSettings( const QgsPalLayerSettings& s )
+    : QgsTextRendererSettings( s )
 {
   // copy only permanent stuff
 
@@ -465,22 +384,8 @@ QgsPalLayerSettings::QgsPalLayerSettings( const QgsPalLayerSettings& s )
   // text style
   fieldName = s.fieldName;
   isExpression = s.isExpression;
-  textFont = s.textFont;
-  textNamedStyle = s.textNamedStyle;
-  fontSizeInMapUnits = s.fontSizeInMapUnits;
-  fontSizeMapUnitScale = s.fontSizeMapUnitScale;
-  textColor = s.textColor;
-  textTransp = s.textTransp;
-  blendMode = s.blendMode;
-  previewBkgrdColor = s.previewBkgrdColor;
-  // font processing info
-  mTextFontFound = s.mTextFontFound;
-  mTextFontFamily = s.mTextFontFamily;
 
   // text formatting
-  wrapChar = s.wrapChar;
-  multilineHeight = s.multilineHeight;
-  multilineAlign = s.multilineAlign;
   addDirectionSymbol = s.addDirectionSymbol;
   leftDirectionSymbol = s.leftDirectionSymbol;
   rightDirectionSymbol = s.rightDirectionSymbol;
@@ -489,17 +394,6 @@ QgsPalLayerSettings::QgsPalLayerSettings( const QgsPalLayerSettings& s )
   formatNumbers = s.formatNumbers;
   decimals = s.decimals;
   plusSign = s.plusSign;
-
-  // text buffer
-  bufferDraw = s.bufferDraw;
-  bufferSize = s.bufferSize;
-  bufferSizeInMapUnits = s.bufferSizeInMapUnits;
-  bufferSizeMapUnitScale = s.bufferSizeMapUnitScale;
-  bufferColor = s.bufferColor;
-  bufferTransp = s.bufferTransp;
-  bufferNoFill = s.bufferNoFill;
-  bufferJoinStyle = s.bufferJoinStyle;
-  bufferBlendMode = s.bufferBlendMode;
 
   // placement
   placement = s.placement;
@@ -540,57 +434,10 @@ QgsPalLayerSettings::QgsPalLayerSettings( const QgsPalLayerSettings& s )
   maxNumLabels = s.maxNumLabels;
   obstacle = s.obstacle;
 
-  // shape background
-  shapeDraw = s.shapeDraw;
-  shapeType = s.shapeType;
-  shapeSVGFile = s.shapeSVGFile;
-  shapeSizeType = s.shapeSizeType;
-  shapeSize = s.shapeSize;
-  shapeSizeUnits = s.shapeSizeUnits;
-  shapeSizeMapUnitScale = s.shapeSizeMapUnitScale;
-  shapeRotationType = s.shapeRotationType;
-  shapeRotation = s.shapeRotation;
-  shapeOffset = s.shapeOffset;
-  shapeOffsetUnits = s.shapeOffsetUnits;
-  shapeOffsetMapUnitScale = s.shapeOffsetMapUnitScale;
-  shapeRadii = s.shapeRadii;
-  shapeRadiiUnits = s.shapeRadiiUnits;
-  shapeRadiiMapUnitScale = s.shapeRadiiMapUnitScale;
-  shapeFillColor = s.shapeFillColor;
-  shapeBorderColor = s.shapeBorderColor;
-  shapeBorderWidth = s.shapeBorderWidth;
-  shapeBorderWidthUnits = s.shapeBorderWidthUnits;
-  shapeBorderWidthMapUnitScale = s.shapeBorderWidthMapUnitScale;
-  shapeJoinStyle = s.shapeJoinStyle;
-  shapeTransparency = s.shapeTransparency;
-  shapeBlendMode = s.shapeBlendMode;
-
-  // drop shadow
-  shadowDraw = s.shadowDraw;
-  shadowUnder = s.shadowUnder;
-  shadowOffsetAngle = s.shadowOffsetAngle;
-  shadowOffsetDist = s.shadowOffsetDist;
-  shadowOffsetUnits = s.shadowOffsetUnits;
-  shadowOffsetMapUnitScale = s.shadowOffsetMapUnitScale;
-  shadowOffsetGlobal = s.shadowOffsetGlobal;
-  shadowRadius = s.shadowRadius;
-  shadowRadiusUnits = s.shadowRadiusUnits;
-  shadowRadiusMapUnitScale = s.shadowRadiusMapUnitScale;
-  shadowRadiusAlphaOnly = s.shadowRadiusAlphaOnly;
-  shadowTransparency = s.shadowTransparency;
-  shadowScale = s.shadowScale;
-  shadowColor = s.shadowColor;
-  shadowBlendMode = s.shadowBlendMode;
-
   // data defined
   dataDefinedProperties = s.dataDefinedProperties;
   mDataDefinedNames = s.mDataDefinedNames;
 
-  // scale factors
-  vectorScaleFactor = s.vectorScaleFactor;
-  rasterCompressFactor = s.rasterCompressFactor;
-
-  ct = NULL;
   extentGeom = NULL;
   expression = NULL;
 }
@@ -599,8 +446,6 @@ QgsPalLayerSettings::QgsPalLayerSettings( const QgsPalLayerSettings& s )
 QgsPalLayerSettings::~QgsPalLayerSettings()
 {
   // pal layer is deleted internally in PAL
-
-  delete ct;
   delete expression;
   delete extentGeom;
 
@@ -3171,23 +3016,6 @@ int QgsPalLayerSettings::sizeToPixel( double size, const QgsRenderContext& c, Si
   return ( int )( scaleToPixelContext( size, c, unit, rasterfactor, mapUnitScale ) + 0.5 );
 }
 
-double QgsPalLayerSettings::scaleToPixelContext( double size, const QgsRenderContext& c, SizeUnit unit, bool rasterfactor, const QgsMapUnitScale& mapUnitScale ) const
-{
-  // if render context is that of device (i.e. not a scaled map), just return size
-  double mapUnitsPerPixel = mapUnitScale.computeMapUnitsPerPixel( c );
-
-  if ( unit == MapUnits && mapUnitsPerPixel > 0.0 )
-  {
-    size = size / mapUnitsPerPixel * ( rasterfactor ? c.rasterScaleFactor() : 1 );
-  }
-  else // e.g. in points or mm
-  {
-    double ptsTomm = ( unit == Points ? 0.352778 : 1 );
-    size *= ptsTomm * c.scaleFactor() * ( rasterfactor ? c.rasterScaleFactor() : 1 );
-  }
-  return size;
-}
-
 // -------------
 
 QgsPalLabeling::QgsPalLabeling()
@@ -4152,15 +3980,15 @@ void QgsPalLabeling::drawLabeling( QgsRenderContext& context )
 
     if ( tmpLyr.shapeDraw )
     {
-      drawLabel( *it, context, tmpLyr, LabelShape );
+      drawLabel( *it, context, tmpLyr, QgsTextRenderer::ShapePart );
     }
 
     if ( tmpLyr.bufferDraw )
     {
-      drawLabel( *it, context, tmpLyr, LabelBuffer );
+      drawLabel( *it, context, tmpLyr, QgsTextRenderer::LabelPart );
     }
 
-    drawLabel( *it, context, tmpLyr, LabelText );
+    drawLabel( *it, context, tmpLyr, QgsTextRenderer::TextPart );
 
     if ( mResults->mLabelSearchTree )
     {
@@ -4278,10 +4106,9 @@ void QgsPalLabeling::drawLabelCandidateRect( pal::LabelPosition* lp, QPainter* p
     drawLabelCandidateRect( lp->getNextPart(), painter, xform );
 }
 
-void QgsPalLabeling::drawLabel( pal::LabelPosition* label, QgsRenderContext& context, QgsPalLayerSettings& tmpLyr, DrawLabelType drawType, double dpiRatio )
+void QgsPalLabeling::drawLabel( pal::LabelPosition* label, QgsRenderContext& context, QgsPalLayerSettings& tmpLyr, QgsTextRenderer::TextComponentPart drawType, double dpiRatio )
 {
   // NOTE: this is repeatedly called for multi-part labels
-  QPainter* painter = context.painter();
   const QgsMapToPixel* xform = &context.mapToPixel();
 
   QgsLabelComponent component;
@@ -4294,7 +4121,7 @@ void QgsPalLabeling::drawLabel( pal::LabelPosition* label, QgsRenderContext& con
   component.setOrigin( outPt );
   component.setRotation( label->getAlpha() );
 
-  if ( drawType == QgsPalLabeling::LabelShape )
+  if ( drawType == QgsTextRenderer::ShapePart )
   {
     // get rotated label's center point
     QgsPoint centerPt( outPt );
@@ -4314,14 +4141,12 @@ void QgsPalLabeling::drawLabel( pal::LabelPosition* label, QgsRenderContext& con
     component.setCenter( centerPt );
     component.setSize( QgsPoint( label->getWidth(), label->getHeight() ) );
 
-    drawLabelBackground( context, component, tmpLyr );
+    QgsTextRenderer::drawBackgroundPart( context, component, tmpLyr );
   }
 
-  if ( drawType == QgsPalLabeling::LabelBuffer
-       || drawType == QgsPalLabeling::LabelText )
+  if ( drawType == QgsTextRenderer::LabelPart
+       || drawType == QgsTextRenderer::TextPart )
   {
-
-    // TODO: optimize access :)
     QString text = (( QgsPalGeometry* )label->getFeaturePart()->getUserGeometry() )->text();
     QString txt = ( label->getPartId() == -1 ? text : QString( text[label->getPartId()] ) );
     QFontMetricsF* labelfm = (( QgsPalGeometry* )label->getFeaturePart()->getUserGeometry() )->getLabelFontMetrics();
@@ -4377,119 +4202,9 @@ void QgsPalLabeling::drawLabel( pal::LabelPosition* label, QgsRenderContext& con
     }
 
     //QgsDebugMsgLevel( "drawLabel " + txt, 4 );
-
     QStringList multiLineList = txt.split( wrapchr );
-    int lines = multiLineList.size();
 
-    double labelWidest = 0.0;
-    for ( int i = 0; i < lines; ++i )
-    {
-      double labelWidth = labelfm->width( multiLineList.at( i ) );
-      if ( labelWidth > labelWidest )
-      {
-        labelWidest = labelWidth;
-      }
-    }
-
-    double labelHeight = labelfm->ascent() + labelfm->descent(); // ignore +1 for baseline
-    //  double labelHighest = labelfm->height() + ( double )(( lines - 1 ) * labelHeight * tmpLyr.multilineHeight );
-
-    // needed to move bottom of text's descender to within bottom edge of label
-    double ascentOffset = 0.25 * labelfm->ascent(); // labelfm->descent() is not enough
-
-    for ( int i = 0; i < lines; ++i )
-    {
-      painter->save();
-      painter->translate( QPointF( outPt.x(), outPt.y() ) );
-      painter->rotate( -label->getAlpha() * 180 / M_PI );
-
-      // scale down painter: the font size has been multiplied by raster scale factor
-      // to workaround a Qt font scaling bug with small font sizes
-      painter->scale( 1.0 / tmpLyr.rasterCompressFactor, 1.0 / tmpLyr.rasterCompressFactor );
-
-      // figure x offset for horizontal alignment of multiple lines
-      double xMultiLineOffset = 0.0;
-      double labelWidth = labelfm->width( multiLineList.at( i ) );
-      if ( lines > 1 && tmpLyr.multilineAlign != QgsPalLayerSettings::MultiLeft )
-      {
-        double labelWidthDiff = labelWidest - labelWidth;
-        if ( tmpLyr.multilineAlign == QgsPalLayerSettings::MultiCenter )
-        {
-          labelWidthDiff /= 2;
-        }
-        xMultiLineOffset = labelWidthDiff;
-        //QgsDebugMsgLevel( QString( "xMultiLineOffset: %1" ).arg( xMultiLineOffset ), 4 );
-      }
-
-      double yMultiLineOffset = ( lines - 1 - i ) * labelHeight * tmpLyr.multilineHeight;
-      painter->translate( QPointF( xMultiLineOffset, - ascentOffset - yMultiLineOffset ) );
-
-      component.setText( multiLineList.at( i ) );
-      component.setSize( QgsPoint( labelWidth, labelHeight ) );
-      component.setOffset( QgsPoint( 0.0, -ascentOffset ) );
-      component.setRotation( -component.rotation() * 180 / M_PI );
-      component.setRotationOffset( 0.0 );
-
-      if ( drawType == QgsPalLabeling::LabelBuffer )
-      {
-        // draw label's buffer
-        drawLabelBuffer( context, component, tmpLyr );
-      }
-      else
-      {
-        // draw label's text, QPainterPath method
-        QPainterPath path;
-        path.addText( 0, 0, tmpLyr.textFont, component.text() );
-
-        // store text's drawing in QPicture for drop shadow call
-        QPicture textPict;
-        QPainter textp;
-        textp.begin( &textPict );
-        textp.setPen( Qt::NoPen );
-        textp.setBrush( tmpLyr.textColor );
-        textp.drawPath( path );
-        // TODO: why are some font settings lost on drawPicture() when using drawText() inside QPicture?
-        //       e.g. some capitalization options, but not others
-        //textp.setFont( tmpLyr.textFont );
-        //textp.setPen( tmpLyr.textColor );
-        //textp.drawText( 0, 0, component.text() );
-        textp.end();
-
-        if ( tmpLyr.shadowDraw && tmpLyr.shadowUnder == QgsPalLayerSettings::ShadowText )
-        {
-          component.setPicture( &textPict );
-          component.setPictureBuffer( 0.0 ); // no pen width to deal with
-          component.setOrigin( QgsPoint( 0.0, 0.0 ) );
-
-          drawLabelShadow( context, component, tmpLyr );
-        }
-
-        // paint the text
-        if ( context.useAdvancedEffects() )
-        {
-          painter->setCompositionMode( tmpLyr.blendMode );
-        }
-
-        // scale for any print output or image saving @ specific dpi
-        painter->scale( component.dpiRatio(), component.dpiRatio() );
-
-        if ( mDrawOutlineLabels )
-        {
-          // draw outlined text
-          _fixQPictureDPI( painter );
-          painter->drawPicture( 0, 0, textPict );
-        }
-        else
-        {
-          // draw text as text (for SVG and PDF exports)
-          painter->setFont( tmpLyr.textFont );
-          painter->setPen( tmpLyr.textColor );
-          painter->setRenderHint( QPainter::TextAntialiasing );
-          painter->drawText( 0, 0, component.text() );
-        }
-      }
-      painter->restore();
-    }
+    QgsTextRenderer::drawTextPart( outPt, QSizeF(), false, multiLineList, drawType, component, tmpLyr, labelfm, context, mDrawOutlineLabels );
   }
 
   // NOTE: this used to be within above multi-line loop block, at end. (a mistake since 2010? [LS])
@@ -4501,500 +4216,21 @@ void QgsPalLabeling::drawLabelBuffer( QgsRenderContext& context,
                                       QgsLabelComponent component,
                                       const QgsPalLayerSettings& tmpLyr )
 {
-  QPainter* p = context.painter();
-
-  double penSize = tmpLyr.scaleToPixelContext( tmpLyr.bufferSize, context,
-                   ( tmpLyr.bufferSizeInMapUnits ? QgsPalLayerSettings::MapUnits : QgsPalLayerSettings::MM ), true, tmpLyr.bufferSizeMapUnitScale );
-
-  QPainterPath path;
-  path.addText( 0, 0, tmpLyr.textFont, component.text() );
-  QPen pen( tmpLyr.bufferColor );
-  pen.setWidthF( penSize );
-  pen.setJoinStyle( tmpLyr.bufferJoinStyle );
-  QColor tmpColor( tmpLyr.bufferColor );
-  // honor pref for whether to fill buffer interior
-  if ( tmpLyr.bufferNoFill )
-  {
-    tmpColor.setAlpha( 0 );
-  }
-
-  // store buffer's drawing in QPicture for drop shadow call
-  QPicture buffPict;
-  QPainter buffp;
-  buffp.begin( &buffPict );
-  buffp.setPen( pen );
-  buffp.setBrush( tmpColor );
-  buffp.drawPath( path );
-  buffp.end();
-
-  if ( tmpLyr.shadowDraw && tmpLyr.shadowUnder == QgsPalLayerSettings::ShadowBuffer )
-  {
-    component.setOrigin( QgsPoint( 0.0, 0.0 ) );
-    component.setPicture( &buffPict );
-    component.setPictureBuffer( penSize / 2.0 );
-
-    drawLabelShadow( context, component, tmpLyr );
-  }
-
-  p->save();
-  if ( context.useAdvancedEffects() )
-  {
-    p->setCompositionMode( tmpLyr.bufferBlendMode );
-  }
-//  p->setPen( pen );
-//  p->setBrush( tmpColor );
-//  p->drawPath( path );
-
-  // scale for any print output or image saving @ specific dpi
-  p->scale( component.dpiRatio(), component.dpiRatio() );
-  _fixQPictureDPI( p );
-  p->drawPicture( 0, 0, buffPict );
-  p->restore();
+  QgsTextRenderer::drawBufferPart( context, component, tmpLyr );
 }
 
 void QgsPalLabeling::drawLabelBackground( QgsRenderContext& context,
     QgsLabelComponent component,
     const QgsPalLayerSettings& tmpLyr )
 {
-  QPainter* p = context.painter();
-  double labelWidth = component.size().x(), labelHeight = component.size().y();
-  //QgsDebugMsgLevel( QString( "Background label rotation: %1" ).arg( component.rotation() ), 4 );
-
-  // shared calculations between shapes and SVG
-
-  // configure angles, set component rotation and rotationOffset
-  if ( tmpLyr.shapeRotationType != QgsPalLayerSettings::RotationFixed )
-  {
-    component.setRotation( -( component.rotation() * 180 / M_PI ) ); // RotationSync
-    component.setRotationOffset(
-      tmpLyr.shapeRotationType == QgsPalLayerSettings::RotationOffset ? tmpLyr.shapeRotation : 0.0 );
-  }
-  else // RotationFixed
-  {
-    component.setRotation( 0.0 ); // don't use label's rotation
-    component.setRotationOffset( tmpLyr.shapeRotation );
-  }
-
-  // mm to map units conversion factor
-  double mmToMapUnits = tmpLyr.shapeSizeMapUnitScale.computeMapUnitsPerPixel( context ) * context.scaleFactor();
-
-  // TODO: the following label-buffered generated shapes and SVG symbols should be moved into marker symbology classes
-
-  if ( tmpLyr.shapeType == QgsPalLayerSettings::ShapeSVG )
-  {
-    // all calculations done in shapeSizeUnits, which are then passed to symbology class for painting
-
-    if ( tmpLyr.shapeSVGFile.isEmpty() )
-      return;
-
-    double sizeOut = 0.0;
-    // only one size used for SVG sizing/scaling (no use of shapeSize.y() or Y field in gui)
-    if ( tmpLyr.shapeSizeType == QgsPalLayerSettings::SizeFixed )
-    {
-      sizeOut = tmpLyr.shapeSize.x();
-    }
-    else if ( tmpLyr.shapeSizeType == QgsPalLayerSettings::SizeBuffer )
-    {
-      // add buffer to greatest dimension of label
-      if ( labelWidth >= labelHeight )
-        sizeOut = labelWidth;
-      else if ( labelHeight > labelWidth )
-        sizeOut = labelHeight;
-
-      // label size in map units, convert to shapeSizeUnits, if different
-      if ( tmpLyr.shapeSizeUnits == QgsPalLayerSettings::MM )
-      {
-        sizeOut /= mmToMapUnits;
-      }
-
-      // add buffer
-      sizeOut += tmpLyr.shapeSize.x() * 2;
-    }
-
-    // don't bother rendering symbols smaller than 1x1 pixels in size
-    // TODO: add option to not show any svgs under/over a certian size
-    if ( tmpLyr.scaleToPixelContext( sizeOut, context, tmpLyr.shapeSizeUnits, false, tmpLyr.shapeSizeMapUnitScale ) < 1.0 )
-      return;
-
-    QgsStringMap map; // for SVG symbology marker
-    map["name"] = QgsSymbolLayerV2Utils::symbolNameToPath( tmpLyr.shapeSVGFile.trimmed() );
-    map["size"] = QString::number( sizeOut );
-    map["size_unit"] = QgsSymbolLayerV2Utils::encodeOutputUnit(
-                         tmpLyr.shapeSizeUnits == QgsPalLayerSettings::MapUnits ? QgsSymbolV2::MapUnit : QgsSymbolV2::MM );
-    map["angle"] = QString::number( 0.0 ); // angle is handled by this local painter
-
-    // offset is handled by this local painter
-    // TODO: see why the marker renderer doesn't seem to translate offset *after* applying rotation
-    //map["offset"] = QgsSymbolLayerV2Utils::encodePoint( tmpLyr.shapeOffset );
-    //map["offset_unit"] = QgsSymbolLayerV2Utils::encodeOutputUnit(
-    //                       tmpLyr.shapeOffsetUnits == QgsPalLayerSettings::MapUnits ? QgsSymbolV2::MapUnit : QgsSymbolV2::MM );
-
-    map["fill"] = tmpLyr.shapeFillColor.name();
-    map["outline"] = tmpLyr.shapeBorderColor.name();
-    map["outline-width"] = QString::number( tmpLyr.shapeBorderWidth );
-
-    // TODO: fix overriding SVG symbol's border width/units in QgsSvgCache
-    // currently broken, fall back to symbol's
-    //map["outline_width_unit"] = QgsSymbolLayerV2Utils::encodeOutputUnit(
-    //                              tmpLyr.shapeBorderWidthUnits == QgsPalLayerSettings::MapUnits ? QgsSymbolV2::MapUnit : QgsSymbolV2::MM );
-
-    if ( tmpLyr.shadowDraw && tmpLyr.shadowUnder == QgsPalLayerSettings::ShadowShape )
-    {
-      // configure SVG shadow specs
-      QgsStringMap shdwmap( map );
-      shdwmap["fill"] = tmpLyr.shadowColor.name();
-      shdwmap["outline"] = tmpLyr.shadowColor.name();
-      shdwmap["size"] = QString::number( sizeOut * tmpLyr.rasterCompressFactor );
-
-      // store SVG's drawing in QPicture for drop shadow call
-      QPicture svgPict;
-      QPainter svgp;
-      svgp.begin( &svgPict );
-
-      // draw shadow symbol
-
-      // clone current render context map unit/mm conversion factors, but not
-      // other map canvas parameters, then substitute this painter for use in symbology painting
-      // NOTE: this is because the shadow needs to be scaled correctly for output to map canvas,
-      //       but will be created relative to the SVG's computed size, not the current map canvas
-      QgsRenderContext shdwContext;
-      shdwContext.setMapToPixel( context.mapToPixel() );
-      shdwContext.setScaleFactor( context.scaleFactor() );
-      shdwContext.setPainter( &svgp );
-
-      QgsSymbolLayerV2* symShdwL = QgsSvgMarkerSymbolLayerV2::create( shdwmap );
-      QgsSvgMarkerSymbolLayerV2* svgShdwM = static_cast<QgsSvgMarkerSymbolLayerV2*>( symShdwL );
-      QgsSymbolV2RenderContext svgShdwContext( shdwContext, QgsSymbolV2::Mixed,
-          ( 100.0 - ( double )( tmpLyr.shapeTransparency ) ) / 100.0 );
-
-      double svgSize = tmpLyr.scaleToPixelContext( sizeOut, context, tmpLyr.shapeSizeUnits, true, tmpLyr.shapeSizeMapUnitScale );
-      svgShdwM->renderPoint( QPointF( svgSize / 2, -svgSize / 2 ), svgShdwContext );
-      svgp.end();
-
-      component.setPicture( &svgPict );
-      // TODO: when SVG symbol's border width/units is fixed in QgsSvgCache, adjust for it here
-      component.setPictureBuffer( 0.0 );
-
-      component.setSize( QgsPoint( svgSize, svgSize ) );
-      component.setOffset( QgsPoint( 0.0, 0.0 ) );
-
-      // rotate about origin center of SVG
-      p->save();
-      p->translate( component.center().x(), component.center().y() );
-      p->rotate( component.rotation() );
-      p->scale( 1.0 / tmpLyr.rasterCompressFactor, 1.0 / tmpLyr.rasterCompressFactor );
-      double xoff = tmpLyr.scaleToPixelContext( tmpLyr.shapeOffset.x(), context, tmpLyr.shapeOffsetUnits, true, tmpLyr.shapeOffsetMapUnitScale );
-      double yoff = tmpLyr.scaleToPixelContext( tmpLyr.shapeOffset.y(), context, tmpLyr.shapeOffsetUnits, true, tmpLyr.shapeOffsetMapUnitScale );
-      p->translate( QPointF( xoff, yoff ) );
-      p->rotate( component.rotationOffset() );
-      p->translate( -svgSize / 2, svgSize / 2 );
-
-      drawLabelShadow( context, component, tmpLyr );
-      p->restore();
-
-      delete svgShdwM;
-      svgShdwM = 0;
-    }
-
-    // draw the actual symbol
-    QgsSymbolLayerV2* symL = QgsSvgMarkerSymbolLayerV2::create( map );
-    QgsSvgMarkerSymbolLayerV2* svgM = static_cast<QgsSvgMarkerSymbolLayerV2*>( symL );
-    QgsSymbolV2RenderContext svgContext( context, QgsSymbolV2::Mixed,
-                                         ( 100.0 - ( double )( tmpLyr.shapeTransparency ) ) / 100.0 );
-
-    p->save();
-    if ( context.useAdvancedEffects() )
-    {
-      p->setCompositionMode( tmpLyr.shapeBlendMode );
-    }
-    p->translate( component.center().x(), component.center().y() );
-    p->rotate( component.rotation() );
-    double xoff = tmpLyr.scaleToPixelContext( tmpLyr.shapeOffset.x(), context, tmpLyr.shapeOffsetUnits, false, tmpLyr.shapeOffsetMapUnitScale );
-    double yoff = tmpLyr.scaleToPixelContext( tmpLyr.shapeOffset.y(), context, tmpLyr.shapeOffsetUnits, false, tmpLyr.shapeOffsetMapUnitScale );
-    p->translate( QPointF( xoff, yoff ) );
-    p->rotate( component.rotationOffset() );
-    svgM->renderPoint( QPointF( 0, 0 ), svgContext );
-    p->setCompositionMode( QPainter::CompositionMode_SourceOver ); // just to be sure
-    p->restore();
-
-    delete svgM;
-    svgM = 0;
-
-  }
-  else  // Generated Shapes
-  {
-    // all calculations done in shapeSizeUnits
-
-    double w = labelWidth / ( tmpLyr.shapeSizeUnits == QgsPalLayerSettings::MM ? mmToMapUnits : 1 );
-    double h = labelHeight / ( tmpLyr.shapeSizeUnits == QgsPalLayerSettings::MM ? mmToMapUnits : 1 );
-
-    double xsize = tmpLyr.shapeSize.x();
-    double ysize = tmpLyr.shapeSize.y();
-
-    if ( tmpLyr.shapeSizeType == QgsPalLayerSettings::SizeFixed )
-    {
-      w = xsize;
-      h = ysize;
-    }
-    else if ( tmpLyr.shapeSizeType == QgsPalLayerSettings::SizeBuffer )
-    {
-      if ( tmpLyr.shapeType == QgsPalLayerSettings::ShapeSquare )
-      {
-        if ( w > h )
-          h = w;
-        else if ( h > w )
-          w = h;
-      }
-      else if ( tmpLyr.shapeType == QgsPalLayerSettings::ShapeCircle )
-      {
-        // start with label bound by circle
-        h = sqrt( pow( w, 2 ) + pow( h, 2 ) );
-        w = h;
-      }
-      else if ( tmpLyr.shapeType == QgsPalLayerSettings::ShapeEllipse )
-      {
-        // start with label bound by ellipse
-        h = h / sqrt( 2.0 ) * 2;
-        w = w / sqrt( 2.0 ) * 2;
-      }
-
-      w += xsize * 2;
-      h += ysize * 2;
-    }
-
-    // convert everything over to map pixels from here on
-    w = tmpLyr.scaleToPixelContext( w, context, tmpLyr.shapeSizeUnits, true, tmpLyr.shapeSizeMapUnitScale );
-    h = tmpLyr.scaleToPixelContext( h, context, tmpLyr.shapeSizeUnits, true, tmpLyr.shapeSizeMapUnitScale );
-
-    // offsets match those of symbology: -x = left, -y = up
-    QRectF rect( -w / 2.0, - h / 2.0, w, h );
-
-    if ( rect.isNull() )
-      return;
-
-    p->save();
-    p->translate( QPointF( component.center().x(), component.center().y() ) );
-    p->rotate( component.rotation() );
-    double xoff = tmpLyr.scaleToPixelContext( tmpLyr.shapeOffset.x(), context, tmpLyr.shapeOffsetUnits, false, tmpLyr.shapeOffsetMapUnitScale );
-    double yoff = tmpLyr.scaleToPixelContext( tmpLyr.shapeOffset.y(), context, tmpLyr.shapeOffsetUnits, false, tmpLyr.shapeOffsetMapUnitScale );
-    p->translate( QPointF( xoff, yoff ) );
-    p->rotate( component.rotationOffset() );
-
-    double penSize = tmpLyr.scaleToPixelContext( tmpLyr.shapeBorderWidth, context, tmpLyr.shapeBorderWidthUnits, true, tmpLyr.shapeBorderWidthMapUnitScale );
-
-    QPen pen;
-    if ( tmpLyr.shapeBorderWidth > 0 )
-    {
-      pen.setColor( tmpLyr.shapeBorderColor );
-      pen.setWidthF( penSize );
-      if ( tmpLyr.shapeType == QgsPalLayerSettings::ShapeRectangle )
-        pen.setJoinStyle( tmpLyr.shapeJoinStyle );
-    }
-    else
-    {
-      pen = Qt::NoPen;
-    }
-
-    // store painting in QPicture for shadow drawing
-    QPicture shapePict;
-    QPainter shapep;
-    shapep.begin( &shapePict );
-    shapep.setPen( pen );
-    shapep.setBrush( tmpLyr.shapeFillColor );
-
-    if ( tmpLyr.shapeType == QgsPalLayerSettings::ShapeRectangle
-         || tmpLyr.shapeType == QgsPalLayerSettings::ShapeSquare )
-    {
-      if ( tmpLyr.shapeRadiiUnits == QgsPalLayerSettings::Percent )
-      {
-        shapep.drawRoundedRect( rect, tmpLyr.shapeRadii.x(), tmpLyr.shapeRadii.y(), Qt::RelativeSize );
-      }
-      else
-      {
-        double xRadius = tmpLyr.scaleToPixelContext( tmpLyr.shapeRadii.x(), context, tmpLyr.shapeRadiiUnits, true, tmpLyr.shapeRadiiMapUnitScale );
-        double yRadius = tmpLyr.scaleToPixelContext( tmpLyr.shapeRadii.y(), context, tmpLyr.shapeRadiiUnits, true, tmpLyr.shapeRadiiMapUnitScale );
-        shapep.drawRoundedRect( rect, xRadius, yRadius );
-      }
-    }
-    else if ( tmpLyr.shapeType == QgsPalLayerSettings::ShapeEllipse
-              || tmpLyr.shapeType == QgsPalLayerSettings::ShapeCircle )
-    {
-      shapep.drawEllipse( rect );
-    }
-    shapep.end();
-
-    p->scale( 1.0 / tmpLyr.rasterCompressFactor, 1.0 / tmpLyr.rasterCompressFactor );
-
-    if ( tmpLyr.shadowDraw && tmpLyr.shadowUnder == QgsPalLayerSettings::ShadowShape )
-    {
-      component.setPicture( &shapePict );
-      component.setPictureBuffer( penSize / 2.0 );
-
-      component.setSize( QgsPoint( rect.width(), rect.height() ) );
-      component.setOffset( QgsPoint( rect.width() / 2, -rect.height() / 2 ) );
-      drawLabelShadow( context, component, tmpLyr );
-    }
-
-    p->setOpacity(( 100.0 - ( double )( tmpLyr.shapeTransparency ) ) / 100.0 );
-    if ( context.useAdvancedEffects() )
-    {
-      p->setCompositionMode( tmpLyr.shapeBlendMode );
-    }
-
-    // scale for any print output or image saving @ specific dpi
-    p->scale( component.dpiRatio(), component.dpiRatio() );
-    _fixQPictureDPI( p );
-    p->drawPicture( 0, 0, shapePict );
-    p->restore();
-  }
+  QgsTextRenderer::drawBackgroundPart( context, component, tmpLyr );
 }
 
 void QgsPalLabeling::drawLabelShadow( QgsRenderContext& context,
                                       QgsLabelComponent component,
                                       const QgsPalLayerSettings& tmpLyr )
 {
-  // incoming component sizes should be multiplied by rasterCompressFactor, as
-  // this allows shadows to be created at paint device dpi (e.g. high resolution),
-  // then scale device painter by 1.0 / rasterCompressFactor for output
-
-  QPainter* p = context.painter();
-  double componentWidth = component.size().x(), componentHeight = component.size().y();
-  double xOffset = component.offset().x(), yOffset = component.offset().y();
-  double pictbuffer = component.pictureBuffer();
-
-  // generate pixmap representation of label component drawing
-  bool mapUnits = ( tmpLyr.shadowRadiusUnits == QgsPalLayerSettings::MapUnits );
-  double radius = tmpLyr.scaleToPixelContext( tmpLyr.shadowRadius , context, tmpLyr.shadowRadiusUnits, !mapUnits, tmpLyr.shadowRadiusMapUnitScale );
-  radius /= ( mapUnits ? tmpLyr.vectorScaleFactor / component.dpiRatio() : 1 );
-  radius = ( int )( radius + 0.5 );
-
-  // TODO: add labeling gui option to adjust blurBufferClippingScale to minimize pixels, or
-  //       to ensure shadow isn't clipped too tight. (Or, find a better method of buffering)
-  double blurBufferClippingScale = 3.75;
-  int blurbuffer = ( radius > 17 ? 16 : radius ) * blurBufferClippingScale;
-
-  QImage blurImg( componentWidth + ( pictbuffer * 2.0 ) + ( blurbuffer * 2.0 ),
-                  componentHeight + ( pictbuffer * 2.0 ) + ( blurbuffer * 2.0 ),
-                  QImage::Format_ARGB32_Premultiplied );
-
-  // TODO: add labeling gui option to not show any shadows under/over a certian size
-  // keep very small QImages from causing paint device issues, i.e. must be at least > 1
-  int minBlurImgSize = 1;
-  // max limitation on QgsSvgCache is 10,000 for screen, which will probably be reasonable for future caching here, too
-  // 4 x QgsSvgCache limit for output to print/image at higher dpi
-  // TODO: should it be higher, scale with dpi, or have no limit? Needs testing with very large labels rendered at high dpi output
-  int maxBlurImgSize = 40000;
-  if ( blurImg.isNull()
-       || ( blurImg.width() < minBlurImgSize || blurImg.height() < minBlurImgSize )
-       || ( blurImg.width() > maxBlurImgSize || blurImg.height() > maxBlurImgSize ) )
-    return;
-
-  blurImg.fill( QColor( Qt::transparent ).rgba() );
-  QPainter pictp;
-  if ( !pictp.begin( &blurImg ) )
-    return;
-  pictp.setRenderHints( QPainter::Antialiasing | QPainter::SmoothPixmapTransform );
-  QPointF imgOffset( blurbuffer + pictbuffer + xOffset,
-                     blurbuffer + pictbuffer + componentHeight + yOffset );
-
-  pictp.drawPicture( imgOffset,
-                     *component.picture() );
-
-  // overlay shadow color
-  pictp.setCompositionMode( QPainter::CompositionMode_SourceIn );
-  pictp.fillRect( blurImg.rect(), tmpLyr.shadowColor );
-  pictp.end();
-
-  // blur the QImage in-place
-  if ( tmpLyr.shadowRadius > 0.0 && radius > 0 )
-  {
-    QgsSymbolLayerV2Utils::blurImageInPlace( blurImg, blurImg.rect(), radius, tmpLyr.shadowRadiusAlphaOnly );
-  }
-
-  if ( tmpLyr.showingShadowRects ) // engine setting, not per layer
-  {
-    // debug rect for QImage shadow registration and clipping visualization
-    QPainter picti;
-    picti.begin( &blurImg );
-    picti.setBrush( Qt::Dense7Pattern );
-    QPen imgPen( QColor( 0, 0, 255, 255 ) );
-    imgPen.setWidth( 1 );
-    picti.setPen( imgPen );
-    picti.setOpacity( 0.1 );
-    picti.drawRect( 0, 0, blurImg.width(), blurImg.height() );
-    picti.end();
-  }
-
-  double offsetDist = tmpLyr.scaleToPixelContext( tmpLyr.shadowOffsetDist, context, tmpLyr.shadowOffsetUnits, true, tmpLyr.shadowOffsetMapUnitScale );
-  double angleRad = tmpLyr.shadowOffsetAngle * M_PI / 180; // to radians
-  if ( tmpLyr.shadowOffsetGlobal )
-  {
-    // TODO: check for differences in rotation origin and cw/ccw direction,
-    //       when this shadow function is used for something other than labels
-
-    // it's 0-->cw-->360 for labels
-    //QgsDebugMsgLevel( QString( "Shadow aggregated label rotation (degrees): %1" ).arg( component.rotation() + component.rotationOffset() ), 4 );
-    angleRad -= ( component.rotation() * M_PI / 180 + component.rotationOffset() * M_PI / 180 );
-  }
-
-  QPointF transPt( -offsetDist * cos( angleRad + M_PI / 2 ),
-                   -offsetDist * sin( angleRad + M_PI / 2 ) );
-
-  p->save();
-  p->setRenderHints( QPainter::Antialiasing | QPainter::SmoothPixmapTransform );
-  if ( context.useAdvancedEffects() )
-  {
-    p->setCompositionMode( tmpLyr.shadowBlendMode );
-  }
-  p->setOpacity(( 100.0 - ( double )( tmpLyr.shadowTransparency ) ) / 100.0 );
-
-  double scale = ( double )tmpLyr.shadowScale / 100.0;
-  // TODO: scale from center/center, left/center or left/top, instead of default left/bottom?
-  p->scale( scale, scale );
-  if ( component.useOrigin() )
-  {
-    p->translate( component.origin().x(), component.origin().y() );
-  }
-  p->translate( transPt );
-  p->translate( -imgOffset.x(),
-                -imgOffset.y() );
-  p->drawImage( 0, 0, blurImg );
-  p->restore();
-
-  // debug rects
-  if ( tmpLyr.showingShadowRects ) // engine setting, not per layer
-  {
-    // draw debug rect for QImage painting registration
-    p->save();
-    p->setBrush( Qt::NoBrush );
-    QPen imgPen( QColor( 255, 0, 0, 10 ) );
-    imgPen.setWidth( 2 );
-    imgPen.setStyle( Qt::DashLine );
-    p->setPen( imgPen );
-    p->scale( scale, scale );
-    if ( component.useOrigin() )
-    {
-      p->translate( component.origin().x(), component.origin().y() );
-    }
-    p->translate( transPt );
-    p->translate( -imgOffset.x(),
-                  -imgOffset.y() );
-    p->drawRect( 0, 0, blurImg.width(), blurImg.height() );
-    p->restore();
-
-    // draw debug rect for passed in component dimensions
-    p->save();
-    p->setBrush( Qt::NoBrush );
-    QPen componentRectPen( QColor( 0, 255, 0, 70 ) );
-    componentRectPen.setWidth( 1 );
-    if ( component.useOrigin() )
-    {
-      p->translate( component.origin().x(), component.origin().y() );
-    }
-    p->setPen( componentRectPen );
-    p->drawRect( QRect( -xOffset, -componentHeight - yOffset, componentWidth, componentHeight ) );
-    p->restore();
-  }
+  QgsTextRenderer::drawShadowPart( context, component, tmpLyr );
 }
 
 void QgsPalLabeling::loadEngineSettings()
